+++ src/data/components/Popover.js
@@ -1,0 +1,9 @@
+export default {
+  name: 'Popover',
+  supports: {
+    bootstrap3: 'https://getbootstrap.com/docs/3.3/javascript/#popovers',
+    bootstrap4: 'https://getbootstrap.com/docs/4.0/components/popovers/',
+    semantic: 'https://semantic-ui.com/modules/popup.html',
+    spectre: 'https://picturepan2.github.io/spectre/components.html#popovers'
+  }
+};