+++ src/data/components/Tooltip.js
@@ -1,10 +1,12 @@
 export default {
   name: 'Tooltip',
   supports: {
+    amaze: 'http://amazeui.github.io/docs/en/javascript/popover/',
     blaze: 'http://blazecss.com/components/tooltips/',
     bootstrap3: 'https://getbootstrap.com/docs/3.3/javascript/#tooltips',
     bootstrap4: 'https://getbootstrap.com/docs/4.0/components/tooltips/',
     semantic: 'http://semantic-ui.com/modules/popup.html#tooltip',
+    foundation5: 'http://foundation.zurb.com/sites/docs/v/5.5.3/components/tooltips.html',
     foundation6: 'http://foundation.zurb.com/sites/docs/tooltip.html',
     ink: 'http://ink.sapo.pt/javascript-ui/#InkUITooltip_1',
     materialize: 'http://materializecss.com/dialogs.html#tooltip',